requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiohttp-apispec==2.2.3
aiohttp-swagger3==0.10.0
aiosignal==1.3.2
alembic==1.16.2
annotated-types==0.7.0
apispec==3.3.2
asyncpg==0.30.0
attrs==25.3.0
bcrypt==4.3.0
certifi==2025.6.15
cffi==1.17.1
cryptography==45.0.4
dotenv==0.9.9
fastjsonschema==2.19.1
frozenlist==1.7.0
greenlet==3.2.3
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==4.0.0
multidict==6.5.0
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
propcache==0.3.2
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
pytest==8.4.1
python-dotenv==1.1.0
PyYAML==6.0.2
rfc3339-validator==0.1.4
six==1.17.0
SQLAlchemy==2.0.41
typing-inspection==0.4.1
typing_extensions==4.14.0
webargs==5.5.3
yarl==1.20.1
